+++ src/main/kotlin/hamburg/remme/tinygit/domain/LocalBranch.kt
@@ -1,4 +1,4 @@
-package hamburg.remme.tinygit.git
+package hamburg.remme.tinygit.domain
 
 class LocalBranch(val shortRef: String, val commitId: String, val remote: Boolean) {
