+++ src/main/kotlin/hamburg/remme/tinygit/domain/LocalDivergence.kt
@@ -1,4 +1,4 @@
-package hamburg.remme.tinygit.git
+package hamburg.remme.tinygit.domain
 
 class LocalDivergence(val ahead: Int, val behind: Int) {
