+++ src/main/kotlin/hamburg/remme/tinygit/domain/LocalGraph.kt
@@ -1,4 +1,4 @@
-package hamburg.remme.tinygit.git
+package hamburg.remme.tinygit.domain
 
 class LocalGraph(commits: List<LocalCommit>) : Iterable<LocalGraph.Node> {
