+++ src/main/kotlin/hamburg/remme/tinygit/domain/LocalRebase.kt
@@ -1,4 +1,4 @@
-package hamburg.remme.tinygit.git
+package hamburg.remme.tinygit.domain
 
 class LocalRebase(val next: Int, val last: Int) {
