+++ src/main/kotlin/hamburg/remme/tinygit/domain/LocalRepository.kt
@@ -1,4 +1,4 @@
-package hamburg.remme.tinygit.git
+package hamburg.remme.tinygit.domain
 
 import hamburg.remme.tinygit.normalize
 import hamburg.remme.tinygit.shorten
