+++ src/main/kotlin/hamburg/remme/tinygit/domain/service/MergeService.kt
@@ -1,9 +1,9 @@
-package hamburg.remme.tinygit.service
+package hamburg.remme.tinygit.domain.service
 
 import hamburg.remme.tinygit.State
-import hamburg.remme.tinygit.git.LocalFile
-import hamburg.remme.tinygit.git.LocalRepository
-import hamburg.remme.tinygit.git.api.Git
+import hamburg.remme.tinygit.domain.LocalFile
+import hamburg.remme.tinygit.domain.LocalRepository
+import hamburg.remme.tinygit.git.Git
 import javafx.collections.ListChangeListener
 
 object MergeService : Refreshable() {
