+++ src/main/kotlin/hamburg/remme/tinygit/domain/service/Refreshable.kt
@@ -1,7 +1,7 @@
-package hamburg.remme.tinygit.service
+package hamburg.remme.tinygit.domain.service
 
 import hamburg.remme.tinygit.State
-import hamburg.remme.tinygit.git.LocalRepository
+import hamburg.remme.tinygit.domain.LocalRepository
 
 abstract class Refreshable {
