+++ src/main/kotlin/hamburg/remme/tinygit/gui/FileStatusView.kt
@@ -1,6 +1,6 @@
 package hamburg.remme.tinygit.gui
 
-import hamburg.remme.tinygit.git.LocalFile
+import hamburg.remme.tinygit.domain.LocalFile
 import hamburg.remme.tinygit.gui.builder.addClass
 import hamburg.remme.tinygit.gui.component.Icons
 import javafx.collections.ObservableList
