+++ src/main/kotlin/hamburg/remme/tinygit/gui/dialog/SettingsDialog.kt
@@ -3,8 +3,8 @@ package hamburg.remme.tinygit.gui.dialog
 import hamburg.remme.tinygit.State
 import hamburg.remme.tinygit.decrypt
 import hamburg.remme.tinygit.encrypt
-import hamburg.remme.tinygit.git.LocalRepository
-import hamburg.remme.tinygit.git.api.Git
+import hamburg.remme.tinygit.domain.LocalRepository
+import hamburg.remme.tinygit.git.Git
 import hamburg.remme.tinygit.gui.builder.addClass
 import hamburg.remme.tinygit.gui.builder.button
 import hamburg.remme.tinygit.gui.builder.columnSpan
